job-fetching-service/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
dateparser==1.2.0
beautifulsoup4==4.12.3
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
python-multipart==0.0.9
